+++ src/main/java/com/epam/reportportal/utils/MimeTypeDetector.java
@@ -37,7 +37,7 @@ import java.util.Map;
 public class MimeTypeDetector {
 	private static final String UNKNOWN_TYPE = "application/octet-stream";
 	private static final String EXTENSION_DELIMITER = ".";
-	private static final int BYTES_TO_READ_FOR_DETECTION = 20;
+	private static final int BYTES_TO_READ_FOR_DETECTION = 128;
 
 	private static final Map<String, String> ADDITIONAL_EXTENSION_MAPPING = Collections.unmodifiableMap(new HashMap<String, String>() {{
 		put(".properties", "text/plain");
@@ -48,7 +48,7 @@ public class MimeTypeDetector {
 		throw new IllegalStateException("Static only class. No instances should exist for the class!");
 	}
 
-	private static int[] readDetectionBytes(@Nonnull InputStream is) throws IOException {
+	static int[] readDetectionBytes(@Nonnull InputStream is) throws IOException {
 		if (!is.markSupported()) {
 			// Trigger UnsupportedOperationException before reading the stream, no users should get there unless they hack with reflections
 			is.reset();
@@ -66,24 +66,38 @@ public class MimeTypeDetector {
 		return bytes;
 	}
 
+	static boolean isBinary(@Nonnull InputStream is) throws IOException {
+		int[] bytes = readDetectionBytes(is);
+		for (int b : bytes) {
+			if (b == 0) {
+				return true;
+			}
+		}
+		return false;
+	}
+
 	@Nullable
 	static String guessContentTypeFromStream(@Nonnull InputStream is) throws IOException {
 		int[] bytes = readDetectionBytes(is);
 		if (bytes.length >= 8) {
-			if (bytes[0] == 0x89 && bytes[1] == 0x50 && bytes[2] == 0x4e && bytes[3] == 0x47 // 4 bytes break
-					&& bytes[4] == 0x0d && bytes[5] == 0x0a && bytes[6] == 0x1a && bytes[7] == 0x0a) {
+			if (bytes[0] == 0x89 && bytes[1] == 'P' && bytes[2] == 'N' && bytes[3] == 'G' // 4 bytes break
+					&& bytes[4] == '\r' && bytes[5] == '\n' && bytes[6] == 0x1a && bytes[7] == '\n') {
 				return "image/png";
 			}
 		}
 		if (bytes.length >= 4) {
-			if (bytes[0] == 0x50 && bytes[1] == 0x4b && bytes[2] == 0x03 && bytes[3] == 0x04) {
+			if (bytes[0] == 'P' && bytes[1] == 'K' && bytes[2] == 0x03 && bytes[3] == 0x04) {
 				// ZIPs
 				if (bytes.length >= 7 && bytes[4] == 0x14 && bytes[5] == 0x00 && bytes[6] == 0x08) {
 					return "application/java-archive";
 				}
 				return "application/zip";
 			}
-			if (bytes[0] == 0x25 && bytes[1] == 0x50 && bytes[2] == 0x44 && bytes[3] == 0x46) {
+			if (bytes[0] == 'P' && bytes[1] == 'K' && bytes[2] == 0x05 && bytes[3] == 0x06) {
+				// Zero-length ZIP
+				return "application/zip";
+			}
+			if (bytes[0] == '%' && bytes[1] == 'P' && bytes[2] == 'D' && bytes[3] == 'F' && isBinary(is)) {
 				return "application/pdf";
 			}
 			if (bytes[0] == 0xFF && bytes[1] == 0xD8 && bytes[2] == 0xFF) {
@@ -92,6 +106,11 @@ public class MimeTypeDetector {
 					// E0 - JPEG/JFIF; E1 - EXIF; E8 - SPIFF
 					return "image/jpeg";
 				}
+			}
+		}
+		if (bytes.length >= 2) {
+			if (bytes[0] == 'B' && bytes[1] == 'M' && isBinary(is)) {
+				return "image/bmp";
 			}
 		}
 		return null;
@@ -139,6 +158,6 @@ public class MimeTypeDetector {
 				type = detectByExtensionInternal(resourceName);
 			}
 		}
-		return type == null ? UNKNOWN_TYPE : type;
+		return type == null ? isBinary(source.openStream()) ? UNKNOWN_TYPE : "text/plain" : type;
 	}
 }