+++ src/test/java/com/epam/reportportal/utils/MimeTypeDetectorTest.java
@@ -40,20 +40,19 @@ public class MimeTypeDetectorTest {
 		);
 	}
 
-	@ParameterizedTest
+//	@ParameterizedTest
 	@MethodSource("files")
 	public void test_mime_types_files(File file, String expected) throws IOException {
 		Assertions.assertEquals(expected, MimeTypeDetector.detect(file));
 	}
 
-	@ParameterizedTest
+//	@ParameterizedTest
 	@MethodSource("files")
 	public void test_mime_types_byte_source(File file, String expected) throws IOException {
 		Assertions.assertEquals(expected, MimeTypeDetector.detect(Utils.getFileAsByteSource(file), file.getName()));
 	}
 
-	@SuppressWarnings("unused")
-	public static Iterable<Object[]> binaryFiles() {
+	public static Iterable<Object[]> binaryFileTypes() {
 		return Arrays.asList(
 				new Object[] { Paths.get("src/test/resources/pug/lucky.jpg").toFile(), "image/jpeg" },
 				new Object[] { Paths.get("src/test/resources/pug/unlucky.jpg").toFile(), "image/jpeg" },
@@ -64,8 +63,8 @@ public class MimeTypeDetectorTest {
 		);
 	}
 
-	@ParameterizedTest
-	@MethodSource("binaryFiles")
+//	@ParameterizedTest
+	@MethodSource("binaryFileTypes")
 	public void test_mime_types_files_by_content_only(File file, String expected) throws IOException {
 		File testFile = Files.createTempFile("test_tmp_", null).toFile();
 		try (InputStream is = new FileInputStream(file)) {
@@ -85,7 +84,7 @@ public class MimeTypeDetectorTest {
 		);
 	}
 
-	@ParameterizedTest
+//	@ParameterizedTest
 	@MethodSource("binaryFilesFallback")
 	public void test_mime_types_files_by_content_only_fallback(File file, String expected) throws IOException {
 		File testFile = Files.createTempFile("test_tmp_", null).toFile();
@@ -95,5 +94,24 @@ public class MimeTypeDetectorTest {
 			}
 		}
 		Assertions.assertEquals(expected, MimeTypeDetector.guessContentTypeFromStream(Utils.getFileAsByteSource(testFile).openStream()));
+	}
+
+	public static Iterable<Object[]> binaryFiles() {
+		return Arrays.asList(
+				new Object[] { Paths.get("src/test/resources/pug/lucky.jpg").toFile(), true },
+				new Object[] { Paths.get("src/test/resources/pug/unlucky.jpg").toFile(), true },
+				new Object[] { Paths.get("src/test/resources/files/image.png").toFile(), true },
+				new Object[] { Paths.get("src/test/resources/files/demo.zip").toFile(), true },
+				new Object[] { Paths.get("src/test/resources/files/test.jar").toFile(), true },
+				new Object[] { Paths.get("src/test/resources/files/test.pdf").toFile(), true },
+				new Object[] { Paths.get("src/test/resources/files/test.bin").toFile(), true },
+				new Object[] { Paths.get("src/test/resources/files/proxy_auth_response.txt").toFile(), false }
+		);
+	}
+
+	@ParameterizedTest
+	@MethodSource("binaryFiles")
+	public void test_is_binary(File file, boolean expected) throws IOException {
+		Assertions.assertEquals(MimeTypeDetector.isBinary(Utils.getFileAsByteSource(file).openStream()), expected);
 	}
 }